feat(deps): Update bouncycastle from 1.64 to 1.72
The bouncycastle version that is currently used is very old (> 3 years)
and also known to be vulnerable ([CVE-2020-15522][1]). The newer
bouncycastle versions also feature some useful additions, e.g. for
[parsing SSH2 RSA keys][2]. Hence, it makes sense to update to the
latest version.

The bouncycastle variant `jdk15on` that is currently used by the
projects under the TLS-Attacker umbrella is not updated anymore and we
need to switch to `jdk18on`. This should be no problem because all
projects should already support Java 8. Unfortunately, the `artifactId`
change requires touching the POM files for each individual project
though.

[1]: https://cve.mitre.org/cgi-bin/cvename.cgi?name=CVE-2020-15522
[2]: https://github.com/bcgit/bc-java/commit/563773c207ab885bb747b791026bf030b3d6d85d